video/external: Move OpenH264 related stuff into the `openh264` decoder module
